// src/app/modules/core/dmo/model/dmo.ts
export class TBillRediscounting {
  issueDate: string;
  originalFaceValue: string;
  currentFaceValue: string;
  rediscountAmount: string;
  faceValueOfRediscounted: string;
  residualAmount: string;
}